feature/chatroom-backend: created user controller
